requirements.txt
langchain
wikipedia
langchain-community
langchain-google-genai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
murf